Branching 3.2 off trunk at revision r3210

+++ buildscripts/phing/classes/phing/tasks/system/MoveTask.php
@@ -1,6 +1,6 @@
 <?php
 /*
- *  $Id: 01e6d627d455729bfed47d603188d8c56f54d9e5 $
+ *  $Id: MoveTask.php 59 2006-04-28 14:49:47Z mrook $
  *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
  * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
@@ -35,7 +35,7 @@ include_once 'phing/system/io/IOException.php';
  * Source files and directories are only deleted when the file or
  * directory has been copied to the destination successfully.
  *
- * @version $Id$
+ * @version $Revision: 1.8 $
  * @package phing.tasks.system
  */
 class MoveTask extends CopyTask {
@@ -45,58 +45,7 @@ class MoveTask extends CopyTask {
         $this->forceOverwrite = true;
     }
     
-    /**
-     * Validates attributes coming in from XML
-     *
-     * @access  private
-     * @return  void
-     * @throws  BuildException
-     */
-    protected function validateAttributes() {    
-        if ($this->file !== null && $this->file->isDirectory()) {
-            if (($this->destFile !== null && $this->destDir !== null)
-                || ($this->destFile === null && $this->destDir === null)) {
-                    throw new BuildException("One and only one of tofile and todir must be set.");
-            }
-            
-            if ($this->destFile === null)
-            {
-                $this->destFile = new PhingFile($this->destDir, $this->file->getName());
-            }
-            
-            if ($this->destDir === null)
-            {
-                $this->destDir = $this->destFile->getParentFile();
-            }
-            
-            $this->completeDirMap[$this->file->getAbsolutePath()] = $this->destFile->getAbsolutePath();
-            
-            $this->file = null;
-        } else {
-            parent::validateAttributes();
-        }
-    }
-    
     protected function doWork() {
-        if (count($this->completeDirMap) > 0)
-        {
-            foreach ($this->completeDirMap as $from => $to)
-            {
-                $f = new PhingFile($from);
-                $d = new PhingFile($to);
-                
-                $moved = false;
-                try { // try to rename                    
-                    $this->log("Attempting to rename $from to $to", $this->verbosity);
-                    $this->fileUtils->copyFile($f, $d, $this->forceOverwrite, $this->preserveLMT, $this->filterChains, $this->getProject(), $this->mode);
-                    $f->delete(true);
-                    $moved = true;
-                } catch (IOException $ioe) {
-                    $moved = false;
-                    $this->logError("Failed to rename $from to $to: " . $ioe->getMessage());
-                }
-            }
-        }
     
         $copyMapSize = count($this->fileCopyMap);
         if ($copyMapSize > 0) {
@@ -109,18 +58,33 @@ class MoveTask extends CopyTask {
                     continue;
                 }
 
+                $moved = false;
                 $f = new PhingFile($from);
                 $d = new PhingFile($to);
                 
-                try { // try to move
-                    $this->log("Moving $from to $to", $this->verbosity);
-
-                    $this->fileUtils->copyFile($f, $d, $this->forceOverwrite, $this->preserveLMT, $this->filterChains, $this->getProject(), $this->mode);
-
-                    $f->delete();
+                $moved = false;
+                try { // try to rename                    
+                    $this->log("Attempting to rename $from to $to", $this->verbosity);
+                    $this->renameFile($f, $d, $this->forceOverwrite);
+                    $moved = true;
                 } catch (IOException $ioe) {
-                    $this->logError("Failed to move $from to $to: " . $ioe->getMessage(), $this->location);
+                    $moved = false;
+                    $this->log("Failed to rename $from to $to: " . $ioe->getMessage(), $this->verbosity);
                 }
+
+                if (!$moved) {                    
+                    try { // try to move
+                        $this->log("Moving $from to $to", $this->verbosity);
+
+                        $this->fileUtils->copyFile($f, $d, $this->forceOverwrite, $this->preserveLMT, $this->filterChains, $this->getProject());                        
+
+                        $f = new PhingFile($fromFile);
+                        $f->delete();
+                    } catch (IOException $ioe) {
+                        $msg = "Failed to move $from to $to: " . $ioe->getMessage();
+                        throw new BuildException($msg, $this->location);
+                    }
+                } // if !moved
             } // foreach fileCopyMap
         } // if copyMapSize
 
@@ -132,7 +96,7 @@ class MoveTask extends CopyTask {
                 $d = new PhingFile((string) $dir);
                 if (!$d->exists()) {
                     if (!$d->mkdirs()) {
-                        $this->logError("Unable to create directory " . $d->getAbsolutePath());
+                        $this->log("Unable to create directory " . $d->getAbsolutePath(), PROJECT_MSG_ERR);
                     } else {
                         $count++;
                     }
@@ -196,7 +160,38 @@ class MoveTask extends CopyTask {
         try {
             $d->delete();
         } catch (Exception $e) {
-            $this->logError("Unable to delete directory " . $d->__toString() . ": " . $e->getMessage());
+            throw new BuildException("Unable to delete directory " . $d->__toString() . ": " . $e->getMessage());
         }
     }
+
+    /**
+     * Attempts to rename a file from a source to a destination.
+     * If overwrite is set to true, this method overwrites existing file
+     * even if the destination file is newer.
+     * Otherwise, the source f
+     * ile is renamed only if the destination file #
+     * is older than it.
+     */
+    private function renameFile(PhingFile $sourceFile, PhingFile $destFile, $overwrite) {
+        $renamed = true;
+
+        // ensure that parent dir of dest file exists!
+        $parent = $destFile->getParentFile();
+        if ($parent !== null) {
+            if (!$parent->exists()) {
+                $parent->mkdirs();
+            }
+        }
+        if ($destFile->exists()) {
+            try {
+                $destFile->delete();
+            } catch (Exception $e) {
+                throw new BuildException("Unable to remove existing file " . $destFile->__toString() . ": " . $e->getMessage());
+            }
+        }
+        $renamed = $sourceFile->renameTo($destFile);
+
+        return $renamed;
+    }
 }
+?>
